Move ReindexCouchViews into corehq.preindex.django_migrations
