+++ corehq/preindex/django_migrations.py
@@ -5,6 +5,9 @@ from pathlib import Path
 
 from couchdbkit.designer import document
 
+from django.core.management import call_command
+from django.db.migrations import RunPython
+
 from .accessors import get_preindex_designs
 
 LOCK_PATH = Path(__file__).parent / "couchviews.lock"
@@ -27,6 +30,30 @@ LOCK_PREAMBLE = """\
 # result of a bug in corehq.couchapps.
 #
 """
+
+
+class ReindexCouchViews(RunPython):
+    """Reindex Couch DB views and Elasticsearch indexes
+
+    Use as needed whenever Couch views are changed and need to be
+    reindexed. This operation may take a long time, depending on what is
+    being reindexed.
+
+    It is planned for the Elasticsearch index building features to move
+    to migration operations dedicated to that purpose alone.
+    See corehq/apps/es/migration_operations.py
+    """
+
+    def __init__(self):
+        super().__init__(self.run, RunPython.noop)
+
+    def deconstruct(self):
+        return (self.__class__.__qualname__, [], {})
+
+    def run(self, apps, schema_editor):
+        call_command("preindex_everything", 8)
+        call_command("sync_finish_couchdb_hq")
+        call_command("ptop_es_manage", flip_all_aliases=True)
 
 
 def iter_couch_lock_lines(preamble=LOCK_PREAMBLE):
